Remove delete empty room from webapp

+++ src/utils/leave-behaviour.ts
@@ -177,22 +177,19 @@ export async function leaveRoomBehaviour(
         }
     }
 
+    /*
     // watcha+ SITIV+
     const roomAfterLeave = matrixClient.getRoom(roomId);
-    console.log(`roomAfterLeave ${roomId}`);
-    if (roomAfterLeave) {
-        const joinedMembers = roomAfterLeave.currentState.getJoinedMemberCount();
-        console.log(`roomAfterLeave - joinedMembers ${joinedMembers}`);
-        if (joinedMembers === 0) {
-            console.log(`Suppression du salon vide ${roomId}`);
-            try {
-                await matrixClient.deleteEmptyRoom(roomId);
-            } catch (err) {
-                console.error("Erreur lors de la suppression du salon vide :", err);
-            }
+    if (lastUser && roomAfterLeave) {
+        console.log(`Suppression du salon vide ${roomId}`);
+        try {
+            await matrixClient.deleteEmptyRoom(roomId);
+        } catch (err) {
+            console.error("Erreur lors de la suppression du salon vide :", err);
         }
     }
     // +watcha
+    */
 }
 
 export const leaveSpace = (space: Room): void => {
